fix(aside sections): rename left and right to correctly reflect
