+++ src/components/layout/DesktopSideLayouts/RightDesktopLayout.tsx
@@ -1,69 +1,94 @@
-import { FaXTwitter } from "react-icons/fa6";
-import DrawerNavigationPair from "../../modal/drawer/DrawerNavigationPair";
-import { HeroIcon } from "../../ui/HeroIcon";
-import { useModal } from "../../../context/GlobalState/ModalProvider";
+import { useNavigate } from "react-router-dom";
 import { useCurrentUser } from "../../../context/Auth/CurrentUserProvider";
-import { useLocation } from "react-router-dom";
-import { useUnseenNotificationIds } from "../../../lib/hooks/mutations/useSeenNotifications";
+import { useTopFiveUsers } from "../../../lib/hooks/useTopFiveUsers";
+import SignupView from "../../entry/SignupView";
+import { TermsAndConditions } from "../../entry/TermsAndConditions";
+import { UseTempAccountButton } from "../../entry/UseTempAccountButton";
+import { GoogleAuthButton } from "../../ui/GoogleAuthButton";
+import { HorizontalStripedText } from "../../ui/HorizontalStripedText";
+import { UserSearchResult } from "../pages/UserSearchResult";
+import { AsideContainer } from "./AsideContainer";
 
 export function RightDesktopLayout () {
 
-    const {setModalType} = useModal();
-    const {currentUser} = useCurrentUser();
-    const location = useLocation();
-    const { data: unseenIds = [] } = useUnseenNotificationIds(currentUser?.id);
+    const { currentUser } = useCurrentUser();
 
+    const {data: topUsers} = useTopFiveUsers();
 
+    const navigate = useNavigate();
 
     return (
-        <div className="hidden lg:flex justify-end lg:w-2/3 py-3">
+        <div className="hidden lg:flex lg:flex-col gap-4 px-10 md:items-start py-3 lg:w-2/3">
 
-        <div className="lg:flex lg:flex-col lg:w-2/3 py-3">
+        {!currentUser && (
+            <AsideContainer>
+                <p className="text-xl font-bold">New to X?</p>
+                <p className="text-twitterTextAlt text-xs">Sign up now to get your own personalized timeline!</p>
 
 
-            <FaXTwitter className="text-white text-4xl"/>
-            
-            <div className="flex w-full h-fit flex-col gap-2 py-2">
-
-                <DrawerNavigationPair name={"Home"} routePath="/">
-                    <HeroIcon iconName="HomeIcon" className="h-7 w-7" solid={location.pathname === `/`}/>
-                </DrawerNavigationPair>
-
-                <DrawerNavigationPair name={"Explore"} routePath={`/explore`}>
-                    <HeroIcon iconName="MagnifyingGlassIcon" className="h-7 w-7" solid={location.pathname === `/explore`}/>
-                </DrawerNavigationPair>
-
-                <DrawerNavigationPair name={"Notifications"} routePath="/notifications">
-                    <HeroIcon iconName="BellIcon" className="h-7 w-7" solid={location.pathname === `/notifications`}/>
-                    {currentUser && unseenIds.length > 0 && (
-                        <div className="w-4 rounded-full h-4 absolute left-3 z-40 bg-(--color-main) top-3 right-9">  </div>
-                    )}
-                </DrawerNavigationPair>
-
-                <DrawerNavigationPair name={"Bookmarks"} routePath="/bookmarks">
-                    <HeroIcon iconName="BookmarkIcon" className="h-7 w-7" solid={location.pathname === `/bookmarks`}/>
-                </DrawerNavigationPair>
-
-                <DrawerNavigationPair name={"About"} routePath="/about">
-                    <HeroIcon iconName="QuestionMarkCircleIcon" className="h-7 w-7" solid={false}/>
-                </DrawerNavigationPair>
-
-                <div onClick={() => setModalType("changeColor")}>
-                    <DrawerNavigationPair name={"Set Theme"} >
-                    <HeroIcon iconName="PaintBrushIcon" className="h-7 w-7" solid={false}/>
-                    </DrawerNavigationPair>
+                <div className="w-full flex flex-col gap-4 pt-4">
+                    <GoogleAuthButton>
+                        Sign up with Google
+                    </GoogleAuthButton>
+                    <HorizontalStripedText>
+                        OR
+                    </HorizontalStripedText>
+                    <UseTempAccountButton invertColor={true}/>
+                    <TermsAndConditions/>
                 </div>
+            </AsideContainer>
+        )}
 
-                <DrawerNavigationPair name={"Profile"} routePath={`/profile/${currentUser?.id}`}>
-                    <HeroIcon iconName="UserIcon" className="h-7 w-7" solid={location.pathname === `/profile/${currentUser?.id}`}/>
-                </DrawerNavigationPair>
-
-                <DrawerNavigationPair name={"Log Out"}>
-                    <HeroIcon iconName="PowerIcon" className="h-7 w-7" solid={false}/>
-                </DrawerNavigationPair>
-
+        {currentUser && (
+            <>
+            <AsideContainer>
+                <p className="text-xl pl-2 font-bold">You might like</p>
+                <div className="w-full flex flex-col pt-4">
+                    {topUsers?.map((id) => (
+                        <UserSearchResult userId={id}/>
+                    ))}
                 </div>
-            </div>
+                <p className="pl-2 text-(--color-main) hover:cursor-pointer" onClick={() => navigate("/explore")}>Show more</p>
+            </AsideContainer>
+
+            <AsideContainer disabled={true}>
+                <p className="text-xl font-bold">What's Happening?</p>
+                <div className="w-full flex flex-col gap-4 pt-4">
+                    <div className="w-full flex">
+                        <div className="w-full flex flex-col">
+                            <p className="text-twitterBorder">Trending in the Netherlands</p>
+                            <p className="font-bold">SouthAfricaAwaits</p>
+                        </div>
+                    </div>
+                    <div className="w-full flex">
+                        <div className="w-full flex flex-col">
+                            <p className="text-twitterBorder">Trending in the Netherlands</p>
+                            <p className="font-bold">Mark Rutte</p>
+                        </div>
+                    </div>
+                    <div className="w-full flex">
+                        <div className="w-full flex flex-col">
+                            <p className="text-twitterBorder">Trending in the Netherlands</p>
+                            <p className="font-bold">Bitcoin</p>
+                        </div>
+                    </div>
+                    <div className="w-full flex">
+                        <div className="w-full flex flex-col">
+                            <p className="text-twitterBorder">Trending in the Netherlands</p>
+                            <p className="font-bold">SouthAfricaAwaits</p>
+                        </div>
+                    </div>
+                    <div className="w-full flex">
+                        <div className="w-full flex flex-col">
+                            <p className="text-twitterBorder">Trending in the Netherlands</p>
+                            <p className="font-bold">SouthAfricaAwaits</p>
+                        </div>
+                    </div>
+                </div>
+            </AsideContainer>
+            </>
+        )}
+
         </div>
     )
